Switch from interval timer to date timer
Interval timer was not counting when you switch tabs
which is no bueno. Date timer will be set up so that
we will also have to be counting start time,
current time, the difference between the two, and
the time that was passed between when stop and start
was clicked, otherwise known as 'paused time'.