Timer focus is now the 'doing'
Timer will connect to the current 'doing' task, watching its time
as it builds.

Should the focus feature just be used to display it's current time?
If a task is already in the completed, does it make sense to add time to
it while it is in there? You could just move it to the "doing" because
semantically, it makes more sense. The focus does make things easier but
it logically doesn't make sense to be doing a task that is already
complete which is why re-labeling it as 'doing' is more sound.